requirements.txt
python-telegram-bot>=20.0
Pillow>=9.0.0
python-dotenv>=1.0.0